+++ netdef/Sources/Sources.py
@@ -6,11 +6,45 @@ from collections import OrderedDict
 SOURCEDICT = OrderedDict()
 
 
-def register(name):
+def register(name, classref=None):
+    """
+    A decorator to register sources. Example::
+
+        from netdef.Sources import BaseSource, Sources
+
+        @Sources.register("NewSourceTemplate")
+        class NewSourceTemplate(BaseSource.BaseSource):
+            def __init__(self, name, shared):
+                ...
+    
+    Can also be called as a normal function::
+
+        from netdef.Sources import BaseSource, Sources
+
+        def setup(shared):
+            Sources.register("NewSourceTemplate", NewSourceTemplate)
+
+        class NewSourceTemplate(BaseSource.BaseSource):
+            def __init__(self, name, shared):
+                ...
+
+    :param str name: Name of the source class
+    :param object classref: Should be *None* if used as a decorator and
+                            a *class* if called as a function
+    
+    :return: A callable that returns a *class* if used as a decorator
+             and a *class* if called as a normal function
+
+    """
     def classdecorator(name, cls):
         SOURCEDICT[name] = cls
         return cls
 
+    #if register is called as a normal function with two arguments
+    if not classref is None:
+        return classdecorator(name, classref)
+    
+    #if register is called as a decorator with one argument
     return functools.partial(classdecorator, name)
 
 
@@ -39,10 +73,13 @@ class Sources:
             for name, activate in activate_sources.items():
                 if int(activate) and not name in added:
                     try:
-                        importlib.import_module(
+                        _mod = importlib.import_module(
                             "{}.Sources.{}".format(base_package, name)
                         )
                         added.append(name)
+                        if hasattr(_mod, "setup"):
+                            getattr(_mod, "setup")(self.shared)
+
                     except ImportError as e:
                         if isinstance(e.name, str):
                             if not e.name.startswith(base_package + ".Sources"):